ride session corrected accept button from chat and offer working done chat corrected
