+++ .idea/misc.xml
@@ -77,7 +77,7 @@
         <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/fragment_rider_request_list.xml" value="0.29393115942028986" />
         <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/gender_dropdown_list_item.xml" value="0.29393115942028986" />
         <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/offer_message_from.xml" value="0.18958333333333333" />
-        <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/offer_message_to.xml" value="0.18958333333333333" />
+        <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/offer_message_to.xml" value="0.536" />
         <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/orange_to_offer_msg.xml" value="0.18958333333333333" />
         <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/popup_send_offer_layout.xml" value="0.23958333333333334" />
         <entry key="..\:/android studio projects/SWIFT/app/src/main/res/layout/request_list_item.xml" value="0.36666666666666664" />
