+++ app/src/main/java/com/example/swift/frontEnd/rider/offers/OfferListAdapter.kt
@@ -28,7 +28,7 @@ import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.database.ValueEventListener
 
-class OfferListAdapter (var context:Context, private val driversOfferList:  ArrayList<DriverOffer>) : RecyclerView.Adapter<OfferListAdapter.ViewHolder>() {
+class OfferListAdapter (private val driversOfferList:  ArrayList<DriverOffer>) : RecyclerView.Adapter<OfferListAdapter.ViewHolder>() {
 
     companion object {
         val OFFER_KEY = "OFFER_KEY"
@@ -61,7 +61,7 @@ class OfferListAdapter (var context:Context, private val driversOfferList:  Arra
                 intent.putExtra(OFFER_KEY, offer)
                 intent.putExtra(IS_RIDER, true)
 
-                context.startActivity(intent)
+                view.context.startActivity(intent)
 
 
             }
@@ -83,135 +83,13 @@ class OfferListAdapter (var context:Context, private val driversOfferList:  Arra
             }
 
             acceptBtn.setOnClickListener{
-                //TODO: remove other offers and Request as well
-
-                var db = FirebaseDatabase.getInstance().reference.child("RideRequests")
-
-
+                val obj = AcceptOfferAction()
                 val position: Int = adapterPosition
-                var pickUpLocation:MutableMap<String, Any>
-                var dropOffLocation:MutableMap<String, Any>
 
-                db.get().addOnSuccessListener {
-                   for (child in it.children) {
-                       val req = child.getValue(RideRequest::class.java)
-                       if (req!!.driverId == driversOfferList[position].driverId) {
-                           pickUpLocation = req.sourceLocation!!
-                           dropOffLocation = req.destinationLocation!!
-
-                           val latLng = LatLng(Common.driversFound[driversOfferList[position].driverId]!!.geoLocation!!.latitude,
-                               Common.driversFound[driversOfferList[position].driverId]!!.geoLocation!!.longitude)
-
-                           val d = FirebaseDatabase.getInstance().getReference("RiderOffers")
-                           var money = ""
-                           d.get().addOnSuccessListener { i ->
-                               for (ch in i.children){
-                                   val offer = ch.getValue(DriverOffer::class.java)
-                                   if (req.driverId == offer?.driverId){
-                                       money = offer!!.text
-                                       var session: RideSession = RideSession( offerId = driversOfferList[position].offerId,
-                                           driverId = driversOfferList[position].driverId,
-                                           riderId = driversOfferList[position].riderId,
-                                           rideState = "Picking_Up",
-                                           pickUpLocation = pickUpLocation,
-                                           dropOffLocation = dropOffLocation,
-                                           driverLat = latLng.latitude,
-                                           driverLng = latLng.longitude,
-                                           vehicleType = req.vehicleType,
-                                           money = money)
-
-
-                                       db = FirebaseDatabase.getInstance().getReference("RideSessions")
-                                       val offerId = db.push()
-                                       session.offerId = offerId.key.toString()
-                                       offerId.setValue(session)
-
-                                       val intent = Intent(context, RiderRideSessionActivity::class.java)
-                                       intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                                       context.startActivity(intent)
-                                       break
-                                   }
-                               }
-                               db = FirebaseDatabase.getInstance().reference
-                               val query1 = db.child("RideRequests")
-                                   .orderByChild("riderId")
-                                   .equalTo(FirebaseAuth.getInstance().currentUser!!.uid)
-
-                               query1.addListenerForSingleValueEvent(object : ValueEventListener {
-                                   override fun onDataChange(dataSnapshot: DataSnapshot) {
-                                       for (docs in dataSnapshot.children) {
-                                           docs.ref.removeValue()
-                                       }
-
-                                   }
-
-                                   override fun onCancelled(databaseError: DatabaseError) {
-                                       Log.e(ContentValues.TAG, "onCancelled", databaseError.toException())
-                                   }
-                               })
-
-                               //removing existing offers for current rider
-                               val query2 = db.child("RiderOffers")
-                                   .orderByChild("riderId")
-                                   .equalTo(FirebaseAuth.getInstance().currentUser!!.uid)
-
-                               query2.addListenerForSingleValueEvent(object : ValueEventListener {
-                                   override fun onDataChange(dataSnapshot: DataSnapshot) {
-                                       for (docs in dataSnapshot.children) {
-                                           docs.ref.removeValue()
-                                       }
-
-                                   }
-
-                                   override fun onCancelled(databaseError: DatabaseError) {
-                                       Log.e(ContentValues.TAG, "onCancelled", databaseError.toException())
-                                   }
-                               })
-                           }
-
-                       }
-                   }
-
-                }
-                // deleting all the requests and offers once the offer is accepted
-                val db1 = FirebaseDatabase.getInstance()
-                RiderSession.getCurrentUser { rider ->
-                    //removing existing request for all drivers
-                    val query1 = db1.reference.child("RideRequests")
-                        .orderByChild("riderId")
-                        .equalTo(rider.riderId)
-
-                    query1.addListenerForSingleValueEvent(object : ValueEventListener {
-                        override fun onDataChange(dataSnapshot: DataSnapshot) {
-                            for (docs in dataSnapshot.children) {
-                                docs.ref.removeValue()
-                            }
-
-                        }
-
-                        override fun onCancelled(databaseError: DatabaseError) {
-                            Log.e(ContentValues.TAG, "onCancelled", databaseError.toException())
-                        }
-                    })
-
-                    //removing existing offers for current rider
-                    val query2 = db1.reference.child("RiderOffers")
-                        .orderByChild("riderId")
-                        .equalTo(rider.riderId)
-
-                    query2.addListenerForSingleValueEvent(object : ValueEventListener {
-                        override fun onDataChange(dataSnapshot: DataSnapshot) {
-                            for (docs in dataSnapshot.children) {
-                                docs.ref.removeValue()
-                            }
-
-                        }
-
-                        override fun onCancelled(databaseError: DatabaseError) {
-                            Log.e(ContentValues.TAG, "onCancelled", databaseError.toException())
-                        }
-                    })
-                }
+                val offerDriverID = driversOfferList[position].driverId
+                val offerRiderID = driversOfferList[position].riderId
+                val offerID = driversOfferList[position].offerId
+                obj.acceptOffer(offerDriverID,offerRiderID,offerID,  view.context)
             }
 
         }
