+++ app/src/main/res/layout/offer_message_to.xml
@@ -24,29 +24,28 @@
 
         <TextView
             android:id="@+id/editTextTextPersonName"
-            android:layout_width="87dp"
-            android:layout_height="48dp"
+            android:layout_width="44dp"
+            android:layout_height="21dp"
             android:layout_marginStart="16dp"
             android:layout_marginTop="8dp"
             android:ems="10"
             android:inputType="textPersonName"
             android:text="Offer"
-            android:textSize="25sp"
+            android:textSize="14sp"
             android:textStyle="bold"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
         <TextView
             android:id="@+id/to_offer_msg_textView"
-            android:layout_width="107dp"
-            android:layout_height="54dp"
-            android:layout_marginStart="16dp"
+            android:layout_width="95dp"
+            android:layout_height="31dp"
             android:ems="10"
             android:gravity="center"
             android:inputType="textPersonName"
             android:text="100"
             android:textColor="@color/black"
-            android:textSize="34sp"
+            android:textSize="20sp"
             android:textStyle="bold"
             app:layout_constraintStart_toStartOf="@+id/editTextTextPersonName"
             app:layout_constraintTop_toBottomOf="@+id/editTextTextPersonName" />
@@ -57,16 +56,21 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
 
+            android:layout_marginStart="3dp"
             android:layout_marginTop="8dp"
-            android:layout_marginEnd="32dp"
-            android:backgroundTint="@color/white"
+            android:layout_marginEnd="16dp"
+            android:layout_marginBottom="8dp"
+            android:backgroundTint="@color/quantum_deeporange50"
             android:text="Accept"
             android:textColor="@color/black"
-            android:textSize="24sp"
+            android:textSize="10sp"
             app:cornerRadius="50dp"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="@+id/to_offer_msg_textView"
-            app:layout_constraintTop_toBottomOf="@+id/to_offer_msg_textView" />
+            app:layout_constraintTop_toBottomOf="@+id/to_offer_msg_textView"
+            app:layout_constraintVertical_bias="0.0" />
     </androidx.constraintlayout.widget.ConstraintLayout>
