+++ app/src/main/res/layout/request_list_item.xml
@@ -40,25 +40,31 @@
 
     <TextView
         android:id="@+id/rideRequest_sourceLocation"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_width="273dp"
+        android:layout_height="27dp"
         android:layout_marginTop="16dp"
         android:maxWidth="250dp"
+        android:maxLines="2"
         android:text="Source Location"
         android:textColor="#CDCDCD"
-        android:textSize="17sp"
+        android:textSize="16sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="@+id/rideRequest_Item_riderName"
         app:layout_constraintTop_toBottomOf="@+id/rideRequest_Item_riderName" />
 
     <TextView
         android:id="@+id/rideRequest_destinationLocation"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_width="273dp"
+        android:layout_height="23dp"
         android:layout_marginTop="16dp"
         android:maxWidth="250dp"
+        android:maxLines="2"
         android:text="Destination Location"
         android:textColor="#CDCDCD"
-        android:textSize="17sp"
+        android:textSize="16sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="@+id/rideRequest_sourceLocation"
         app:layout_constraintTop_toBottomOf="@+id/rideRequest_sourceLocation" />
